699: Moving filter under the fapi pkg
https://github.com/SecureApiGateway/SecureApiGateway/issues/699
