+++ secure-api-gateway-ig-extensions/src/main/java/com/forgerock/sapi/gateway/fapi/v1/FAPIAdvancedDCRValidationFilter.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.forgerock.sapi.gateway.dcr.fapi.v1;
+package com.forgerock.sapi.gateway.fapi.v1;
 
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
